# openapi.yaml
openapi: 3.0.0
info:
  title: Kazi Connect API
  version: 1.2.0
  description: >
    Kazi Connect job listings and job applications for SettleIn students, plus accounts and
    employer tools for the FundiLink website. Endpoint numbers match ENDPOINT_LIST.md, and each
    endpoint names the need statement it serves. Endpoints marked with a lock need a login token:
    log in with Endpoint 11, then send the token as "Authorization: Bearer <token>".
servers:
  - url: http://localhost:3000
    description: Local development server
paths:
...
  /api/jobs/{jobId}/applications:
    get:
      summary: Lists the applicants for a job (the employer who posted it)
      description: >
        FundiLink Statement F3. Newest first. An empty list means nobody has applied yet. Each
        applicant includes studentName when the application came from a Kazi Connect account.
      security:
        - bearerAuth: []
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            example: job_12345
      responses:
        '200':
          description: The job's applicants.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Applicant'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Job ID not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Job job_99999 not found'

  #---ENDPOINT 5---
  /api/applications/{applicationId}/status:
    get:
      summary: Returns the current status of an application/service request
      description: Statement 5.
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            example: app_1001
      responses:
        '200':
          description: Successful response containing the application status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationStatus'
        '404':
          description: Application ID not found.

  #---ENDPOINT 6---
  /api/applications/{applicationId}/subscribe:
    get:
      summary: Opens a WebSocket that pushes status changes in real time (accepted / filled)
      description: >
        Statement 5. This is a plain GET that upgrades to a WebSocket, so connect with a WebSocket
        client (ws://), not a normal HTTP request. On connect the server sends the current status
        once, then sends a new message every time the status changes, for example to accepted or
        filled. Every message is a JSON ApplicationStatus object. If the application is cancelled
        with DELETE, a final message with status "cancelled" is sent and the server closes the
        connection. Swagger UI cannot open WebSockets, so Try it out returns 426 for this endpoint.
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            example: app_1001
      responses:
        '101':
          description: Switching Protocols. The real time stream subscription is established.
        '404':
          description: Application ID not found. The connection is refused.
        '426':
          description: Upgrade Required. The request was plain HTTP, not a WebSocket upgrade.

  #---ENDPOINT 7---
  /api/applications:
    post:
      summary: Creates a new application/service request from a SettleIn student
      description: >
        Statement 6. The job must exist and be active. Each call creates a new application, so this
        endpoint is not idempotent: sending the same body twice creates two applications.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NewApplication'
      responses:
        '201':
          description: Application successfully created. The Location header points to the new application.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: >
            The body failed validation: a required field is missing, a field has the wrong type,
            a value is unusable (empty text, not a real date-time, a date in the past), a field
            is not allowed, or jobId does not match any job. Nothing is written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Missing required field(s): preferredDate'
        '409':
          description: The job exists but is closed, so it is not accepting applications. Nothing is written.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Job job_12350 is closed and is not accepting applications

  #---ENDPOINT 8 and ENDPOINT 9 (same path, different methods)---
  /api/applications/{applicationId}:
    put:
      summary: Updates an existing pending application (e.g. reschedule, edit details)
      description: >
        Statement 7. Replaces the editable details of a pending application. This sets an absolute
        state: leaving note out clears it, and sending the same body twice leaves the application in
        the same state. A PUT to an ID that does not exist returns 404 and never creates one.
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            example: app_1001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationUpdate'
      responses:
        '200':
          description: Application successfully updated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: The body failed validation. Nothing is changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'preferredDate must be a real date-time, for example 2027-01-15T09:00:00Z'
        '404':
          description: Application ID not found. Nothing is created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Application app_9999 not found
        '409':
          description: The application is no longer pending (already accepted or filled), so it cannot be changed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Application app_1002 is accepted; only pending applications can be changed'
    delete:
      summary: Cancels a pending application/service request
      description: >
        Statement 8. Removes a pending application. Afterwards Endpoint 5 returns 404 for it, and
        Endpoint 6 subscribers receive a final message with status "cancelled".
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            example: app_1004
      responses:
        '204':
          description: Application successfully cancelled. No body.
        '404':
          description: Application ID not found, including one that was already cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: Application app_9999 not found
        '409':
          description: The application is no longer pending (already accepted or filled), so it cannot be cancelled.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Application app_1002 is accepted; only pending applications can be cancelled'
    # ENDPOINT 18 (same path as Endpoints 8 and 9)
    patch:
      summary: Accepts an application (the employer who posted the job)
      description: >
        FundiLink Statement F3, and SettleIn Statement 5. The application becomes accepted, every
        other application still waiting for the same job becomes filled, and the job closes.
        Subscribers to each application hear about the change live.
      security:
        - bearerAuth: []
      parameters:
        - name: applicationId
          in: path
          required: true
          schema:
            type: string
            example: app_1001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationDecision'
      responses:
        '200':
          description: Application accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Application'
        '400':
          description: status must be accepted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'status must be accepted'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          description: Application ID not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Application app_9999 not found'
        '409':
          description: The application is no longer pending.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
              example:
                message: 'Application app_1002 is accepted; only pending applications can be accepted'

  #---ENDPOINT 10---
...
  /api/me/applications:
    get:
      summary: Lists the logged in student's applications
      description: >
        FundiLink Statement F5. Applications whose studentId is the student's account ID, newest
        first. Empty list if the student has not applied yet.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: The student's applications.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Application'
        '401':
          $ref: '#/components/responses/NotLoggedIn'
        '403':
          $ref: '#/components/responses/Forbidden'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 'The token from Endpoint 10 or 11, sent as "Authorization: Bearer <token>".'
  responses:
    NotLoggedIn:
      description: No token was sent, or it is not valid, or it has expired.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'Your session has expired or is not valid. Log in again.'
    Forbidden:
      description: Logged in, but this account is not allowed to do this.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            message: 'You can only manage jobs you posted.'
  schemas:
...
    ApplicationStatus:
      type: object
      required: [applicationId, status]
      properties:
        applicationId:
          type: string
          example: app_1001
        status:
          type: string
          enum: [pending, accepted, filled, cancelled]
          description: >
            pending is waiting for a decision, accepted means the request was accepted, filled means
            the job went to someone else, and cancelled only ever appears as the final Endpoint 6
            message after a DELETE.
          example: pending
    Application:
      type: object
      required: [applicationId, jobId, studentId, preferredDate, note, status]
      properties:
        applicationId:
          type: string
          example: app_1005
        jobId:
          type: string
          example: job_12345
        studentId:
          type: string
          example: stu_0042
        preferredDate:
          type: string
          format: date-time
          example: '2027-01-15T09:00:00Z'
        note:
          type: string
          description: Empty string when the student left no note.
          example: Available after 2pm on weekdays
        status:
          type: string
          enum: [pending, accepted, filled]
          example: pending
    NewApplication:
      type: object
      additionalProperties: false
      required: [jobId, studentId, preferredDate]
      properties:
        jobId:
          type: string
          minLength: 1
          description: Must match an existing, active job.
          example: job_12345
        studentId:
          type: string
          minLength: 1
          example: stu_0042
        preferredDate:
          type: string
          format: date-time
          description: When the student wants the service. Must be a real date-time in the future.
          example: '2027-01-15T09:00:00Z'
        note:
          type: string
          maxLength: 500
          example: Available after 2pm on weekdays
    ApplicationUpdate:
      type: object
      additionalProperties: false
      required: [preferredDate]
      properties:
        preferredDate:
          type: string
          format: date-time
          description: The new date and time. Must be a real date-time in the future.
          example: '2027-01-22T14:00:00Z'
        note:
          type: string
          maxLength: 500
          description: Replaces the existing note. Leave it out to clear the note.
          example: Rescheduled to the following week
    Error:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: 'Missing required field(s): preferredDate'
...
    Applicant:
      type: object
      description: An Application, plus the applicant's name for the employer deciding on it.
      required: [applicationId, jobId, studentId, preferredDate, note, status, studentName]
      properties:
        applicationId:
          type: string
          example: app_1005
        jobId:
          type: string
          example: job_12346
        studentId:
          type: string
          example: usr_1003
        preferredDate:
          type: string
          format: date-time
          example: '2027-01-15T09:00:00Z'
        note:
          type: string
          example: Free after lectures
        status:
          type: string
          enum: [pending, accepted, filled]
          example: pending
        studentName:
          type: string
          description: Empty when the application did not come from a Kazi Connect account, for example one SettleIn created.
          example: Amina Otieno
...
    ApplicationDecision:
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          enum: [accepted]
          example: accepted